requirements/base.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
sqlite-utils
sqlalchemy

